[Minor] Move pickup cart command to proper namespace
